requirements.txt
Flask
Flask_PyMongo
openai==0.28
pypdf
bcrypt
Flask-Cors==4.0.0
gunicorn
blinker==1.6.2
click==8.1.7
colorama==0.4.6
flask==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
werkzeug==3.0.0
zipp==3.17.0

